+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Update the `org.realityforge.react4j` dependencies to version `0.114`.
+
 ### [v6.24](https://github.com/realityforge/replicant/tree/v6.24) (2019-01-18)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v6.23...v6.24)